+++ hsdm/ChangeLog.md
@@ -3,3 +3,5 @@
 ## 0.1.0.0  -- YYYY-mm-dd
 
 * Basic functions for SDM vector manipulation.
+* IO FFI Query functions
+* Basic vector algebra